Make react default
* clean up package.json
* Update focus test to work with react
