+++ package.json
@@ -42,17 +42,18 @@
     "ajv": "^4.11.2",
     "dot-prop-immutable": "^1.3.1",
     "es6-shim": "^0.35.3",
-    "inferno-component": "^3.10.1",
-    "inferno-create-element": "3.10.1",
-    "inferno-redux": "^3.10.1",
     "json-refs": "^2.1.6",
     "lodash": "^4.17.4",
+    "react": "^16.1.1",
+    "react-dom": "^16.1.1",
+    "react-redux": "^5.0.6",
     "redux": "^3.7.2",
     "redux-thunk": "^2.2.0",
-    "sortablejs": "^1.6.0",
     "uuid": "^3.1.0"
   },
   "devDependencies": {
+    "@types/react": "^16.0.25",
+    "@types/react-redux": "^5.0.13",
     "ava": "^0.18.1",
     "awesome-typescript-loader": "^3.2.0-rc.0",
     "babel": "^6.23.0",
@@ -75,11 +76,8 @@
     "coveralls": "^2.13.3",
     "cross-env": "^5.0.5",
     "document-register-element": "^1.4.1",
-    "drag-mock": "^1.4.0",
     "file-loader": "^0.10.0",
     "html-webpack-plugin": "^2.28.0",
-    "incompose": "^1.1.0",
-    "inferno-test-utils": "^3.9.0",
     "jquery": "^3.2.1",
     "jsdom": "^9.11.0",
     "jsdom-global": "2.1.1",
@@ -87,6 +85,7 @@
     "materialize-css": "0.100.2",
     "ncp": "^2.0.0",
     "nyc": "^10.1.2",
+    "react-addons-test-utils": "^15.6.2",
     "source-map-loader": "^0.2.3",
     "transform-runtime": "0.0.0",
     "ts-loader": "^2.1.0",
