+++ test/helpers/inferno-test.ts
@@ -1,33 +1,32 @@
-// import * as TestUtils from 'react-dom/test-utils';
-import * as TestUtils from 'inferno-test-utils';
-import { Component } from '../../src/common/binding';
-
-export const findRenderedDOMElementWithTag =
-  (component: Component<any, any>, tag: string): Element =>
-    TestUtils.findRenderedDOMElementWithTag(component, tag);
-
-export const findRenderedDOMElementWithClass =
-  (component: Component<any, any>, tag: string): Element =>
-    TestUtils.findRenderedDOMElementWithClass(component, tag);
-
-export const scryRenderedDOMElementsWithTag =
-  (component: Component<any, any>, tag: string): Element[] =>
-    TestUtils.scryRenderedDOMElementsWithTag(component, tag);
-
-export const scryRenderedDOMElementsWithClass =
-  (component: Component<any, any>, tag: string): Element[] =>
-    TestUtils.scryRenderedDOMElementsWithClass(component, tag);
-
-// TODO: figure out types
-export const renderIntoDocument = (element: any): any =>
-  TestUtils.renderIntoDocument(element);
-
-export const click = (button: HTMLElement) => button.click();
-
-export const change = (input: HTMLElement) => {
-  const evt = new Event('change', {
-    'bubbles': true,
-    'cancelable': true
-  });
-  input.dispatchEvent(evt);
-};
+// import * as TestUtils from 'inferno-test-utils';
+// import { Component } from '../../src/common/binding';
+//
+// export const findRenderedDOMElementWithTag =
+//   (component: Component<any, any>, tag: string): Element =>
+//     TestUtils.findRenderedDOMElementWithTag(component, tag);
+//
+// export const findRenderedDOMElementWithClass =
+//   (component: Component<any, any>, tag: string): Element =>
+//     TestUtils.findRenderedDOMElementWithClass(component, tag);
+//
+// export const scryRenderedDOMElementsWithTag =
+//   (component: Component<any, any>, tag: string): Element[] =>
+//     TestUtils.scryRenderedDOMElementsWithTag(component, tag);
+//
+// export const scryRenderedDOMElementsWithClass =
+//   (component: Component<any, any>, tag: string): Element[] =>
+//     TestUtils.scryRenderedDOMElementsWithClass(component, tag);
+//
+// // TODO: figure out types
+// export const renderIntoDocument = (element: any): any =>
+//   TestUtils.renderIntoDocument(element);
+//
+// export const click = (button: HTMLElement) => button.click();
+//
+// export const change = (input: HTMLElement) => {
+//   const evt = new Event('change', {
+//     'bubbles': true,
+//     'cancelable': true
+//   });
+//   input.dispatchEvent(evt);
+// };
