+++ test/renderers/boolean.control.test.tsx
@@ -13,7 +13,8 @@ import {
   change,
   findRenderedDOMElementWithClass,
   findRenderedDOMElementWithTag,
-  renderIntoDocument
+  renderIntoDocument,
+  scryRenderedDOMElementsWithTag
 } from '../helpers/binding';
 import { Provider } from '../../src/common/binding';
 
@@ -89,16 +90,16 @@ test('autofocus on first element', t => {
         schema,
         uischema
     );
-    renderIntoDocument(
+    const tree = renderIntoDocument(
         <Provider store={store}>
             <HorizontalLayoutRenderer schema={schema}
                                       uischema={uischema}
             />
         </Provider>
     );
-    const activeElement = document.activeElement.getElementsByTagName('input')[0].id;
-    t.is(activeElement, '#/properties/firstBooleanField');
-    t.not(activeElement, '#/properties/secondBooleanField');
+    const inputs = scryRenderedDOMElementsWithTag(tree, 'input');
+    t.not(document.activeElement, inputs[0]);
+    t.is(document.activeElement, inputs[1]);
 });
 
 test('autofocus active', t => {
@@ -120,7 +121,7 @@ test('autofocus active', t => {
         </Provider>
     );
     const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
-    t.true(input.autofocus);
+    t.is(document.activeElement, input);
 });
 
 test('autofocus inactive', t => {
@@ -161,7 +162,7 @@ test('autofocus inactive by default', t => {
         </Provider>
     );
     const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
-    t.false(input.autofocus);
+    t.not(document.activeElement, input);
 });
 
 test('tester', t => {
