+++ test/renderers/date.control.test.tsx
@@ -12,7 +12,8 @@ import {
   change,
   findRenderedDOMElementWithClass,
   findRenderedDOMElementWithTag,
-  renderIntoDocument
+  renderIntoDocument,
+  scryRenderedDOMElementsWithTag
 } from '../helpers/binding';
 import { Provider } from '../../src/common/binding';
 
@@ -90,16 +91,16 @@ test('autofocus on first element', t => {
         schema,
         uischema
     );
-    renderIntoDocument(
+    const tree = renderIntoDocument(
         <Provider store={store}>
             <HorizontalLayoutRenderer schema={schema}
                                       uischema={uischema}
             />
         </Provider>
     );
-    const activeElement = document.activeElement.getElementsByTagName('input')[0].id;
-    t.is(activeElement, '#/properties/firstDate');
-    t.not(activeElement, '#/properties/secondDate');
+    const inputs = scryRenderedDOMElementsWithTag(tree, 'input');
+    t.not(document.activeElement, inputs[0]);
+    t.is(document.activeElement, inputs[1]);
 });
 
 test('autofocus active', t => {
@@ -121,7 +122,7 @@ test('autofocus active', t => {
         </Provider>
     );
     const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
-    t.true(input.autofocus);
+    t.is(document.activeElement, input)
 });
 
 test('autofocus inactive', t => {
